+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+## 0.0.7
+
+- feat: listenWhen extension
+- feat: listen extension
+- feat: CubixListener widget
+
+## 0.0.6
+
+- fix: buildWhen extension
+
+## 0.0.5
+
+- fix: cubix extension
+
+## 0.0.4
+
+- chore: support buildWhen
+
+## 0.0.3
+
+- refactor: using RepositoryProvider instead of BlocProvider
+
+## 0.0.2
+
+- docs
+- refactor: CreateContext
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release